added source dir naming changes in main.c key_board_mapping added optimise code with less resource usage
