+++ build.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+# build script for the project will configure cmake leter this is good to go
+
+gcc ./src/*.c -o ./o/tim_kernel.o #./src/*.h
