--- main.c
@@ -1,61 +1,0 @@
-﻿#ifdef HAVE_STDIO_H
-#include <stdio.h>
-#endif
-
-#include <stdio.h>
-
-void setupKeys(){
-	static keyboardKeys[] = {
-		10001, 00001, 10001, 10010, 11001, 
-		101010, 10010, 0.01001, 1000, 10001
-		};
-		
-		char ky1 = keyboardKeys[0];
-		char ky2 = keyboardKeys[1];
-		char ky3 = keyboardKeys[3];
-		char ky4 = keyboardKeys[8];
-		
-		
-struct keyboardMemory{
-	char key;
-	
-	
-	};
-	
-struct keyboardMemory keyLb;
-	keyLb.key = ky1;
-	printf("configuring keyboard USB: [OK] \n");	
-	printf("USB keyboard Addr: %p", keyLb.key);	
-		
-		
-		
-		
-		
-	static usbPortKey1[] = {10001, 00001, 10001, 10010, 11001, 101010, 10010, 0.01001, 1000, 10001 
-		};
-	
-		
-		static usbPortKey2[] = {10001, 00001, 10001, 10010, 11001, 101010, 10010, 0.01001, 1000, 10001 
-		};
-		
-		
-		static bluetoothKeys[] = {10001, 00001, 10001, 10010, 11001, 101010, 10010, 0.01001, 1000, 10001 
-		};
-	char bt1 = bluetoothKeys[0];
-	char bt2 = bluetoothKeys[1];
-	char bt3 = bluetoothKeys[2];
-	char bt4= bluetoothKeys[8];
-	
-	}	
-	
-	
-	
-int main(){
-	setupKeys();
-	return 0;
-	clockid_t;
-	}
-		
-	
-	
-
