+++ run.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+
+./o/tim_kernel.o
