+++ src/tim_kernel.c
@@ -1,0 +1,20 @@
+﻿/*
+ * Main file for this kernel
+ */
+
+#include "tim_kernel.h"
+
+#include <stdio.h>
+
+	
+	
+int main(){
+	setupKeys();
+	return 0;
+	// clockid_t;
+}
+		
+	
+	
+	
+
