+++ src/tim_kernel.h
@@ -1,0 +1,8 @@
+#ifndef  TIM_KERNEL_HEADER
+#define TIM_KERNEL_HEADER
+
+//  functions we need
+
+#include "key_board_mapping.h"
+
+#endif // ! MAIN_HEADER
